backend/requirements.txt

fastapi
uvicorn[standard]
openai
python-dotenv
numpy
requests
httpx
pinecone-client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
slack_sdk
python-multipart
python-jose[cryptography]
